Beyond resolving the mechanical merge conflicts, the AgensGraph fork code was
adapted to PostgreSQL 18's changed internal APIs so that it builds, and the
runtime regressions and incompatibilities exposed by PostgreSQL 18's new
behavior were fixed so that `make check-world` passes cleanly.

Build (adapt AgensGraph fork code to PostgreSQL 18 internal APIs)
- Update the VLE two-sided hash join (nodeHash2Side.c) for PostgreSQL 18's
  reworked hash-join hash-function storage.
- Update the Cypher grammar (gram.y) for the changed parser-helper
  signatures (makeAArrayExpr's end location, mergeTableFuncParameters
  taking the scanner, and related helpers).
- Use the current tuple-descriptor accessor (TupleDescAttr) in the graph
  write executor (execCypherSet.c) in place of the removed direct attrs[]
  access.
- Adapt psql's graph object description (cypher_describe.c, describe.c) and
  the remaining Cypher parser/deparser/executor glue (parse_graph.c,
  ruleutils.c, nodeFuncs.c, graphcmds.c, cypher_funcs.c, execGraphVle.c) to
  the PostgreSQL 18 describe, deparse and node interfaces.

Cypher engine
- Do not parallelize graph-writing statements.  A Cypher write with a
  RETURN clause (CREATE/MERGE/SET/DELETE ... RETURN) is planned as a
  CMD_SELECT whose plan carries a ModifyGraph node, so PostgreSQL 18's
  wider parallel-query eligibility (notably the materialization-aware
  consider_parallel_nestloop) placed a Gather above the ModifyGraph and
  dropped the write entirely -- e.g. "MERGE (a)-[r]-(b) RETURN
  startnode(r)" returned NULL and created no edge.  standard_planner() now
  excludes hasGraphwriteClause queries from parallel mode, mirroring the
  existing hasModifyingCTE exclusion; read-only Cypher still parallelizes.

pg_dump / pg_upgrade of graphs
- Keep the NOT NULL constraint on graph label columns during a
  binary-upgrade CREATE VLABEL/ELABEL.  PostgreSQL 18 records NOT NULL as
  inheritable pg_constraint rows and requires every inheritance child to
  carry one when the parent's primary key is restored, so the previous
  strip-all-constraints path made pg_upgrade fail with
  'column "id" of table "..." is not marked NOT NULL'.
- Allow TRUNCATE on an empty graph label.  Parallel pg_restore issues
  TRUNCATE ... COPY on the tables it just created as a WAL-skipping
  optimization; truncating an empty label is a no-op that cannot orphan
  edges or stale ag_graphmeta, so the label guard now rejects only
  non-empty labels.  This restores pg_dump/pg_restore and pg_upgrade
  round-trips of graphs that contain data.
- Restore the "use File::Compare;" dropped from the pg_upgrade TAP test
  during the merge, which the ag_graphmeta round-trip check depends on.
- Make the regression database self-contained so the new PostgreSQL 18
  dump/restore round-trip test can reload it into a fresh cluster:
  cypher_finalize, propertyindex and cypher_privileges now drop the
  graphs, roles and global default-privilege grants they used to leave
  behind.

Expected-output updates for PostgreSQL 18 planner and display changes
- Adopt the OR-to-ANY index-qual transformation (cypher_expr,
  cypher_graphmeta_prune).
- Show the new "Disabled: true" EXPLAIN marker (cypher_call, cypher_expr).
- Show catalogued NOT NULL constraints in \d output (cypher_ddl).
- Track the shifted parser error position for the CREATE typo test
  (test_extensions).

All of src/test/regress, contrib, src/test/modules, isolation, ecpg,
src/pl and the pg_upgrade TAP suite pass under `make check-world`.

AGV2-442 #done

+++ contrib/pageinspect/hashfuncs.c
@@ -2,7 +2,7 @@
  * hashfuncs.c
  *		Functions to investigate the content of HASH indexes
  *
- * Copyright (c) 2017-2024, PostgreSQL Global Development Group
+ * Copyright (c) 2017-2025, PostgreSQL Global Development Group
  *
  * IDENTIFICATION
  *		contrib/pageinspect/hashfuncs.c
@@ -436,8 +436,8 @@ hash_bitmap_info(PG_FUNCTION_ARGS)
 	if (ovflblkno >= RelationGetNumberOfBlocks(indexRel))
 		ereport(ERROR,
 				(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-				 errmsg("block number %lld is out of range for relation \"%s\"",
-						(long long int) ovflblkno, RelationGetRelationName(indexRel))));
+				 errmsg("block number %" PRId64 " is out of range for relation \"%s\"",
+						ovflblkno, RelationGetRelationName(indexRel))));
 
 	/* Read the metapage so we can determine which bitmap page to use */
 	metabuf = _hash_getbuf(indexRel, HASH_METAPAGE, HASH_READ, LH_META_PAGE);
