+++ src/test/modules/test_slru/test_slru.c
@@ -3,7 +3,7 @@
  * test_slru.c
  *		Test correctness of SLRU functions.
  *
- * Portions Copyright (c) 1996-2024, PostgreSQL Global Development Group
+ * Portions Copyright (c) 1996-2025, PostgreSQL Global Development Group
  * Portions Copyright (c) 1994, Regents of the University of California
  *
  * IDENTIFICATION
@@ -151,8 +151,8 @@ test_slru_page_sync(PG_FUNCTION_ARGS)
 	ftag.segno = pageno / SLRU_PAGES_PER_SEGMENT;
 	SlruSyncFileTag(TestSlruCtl, &ftag, path);
 
-	elog(NOTICE, "Called SlruSyncFileTag() for segment %lld on path %s",
-		 (long long) ftag.segno, path);
+	elog(NOTICE, "Called SlruSyncFileTag() for segment %" PRIu64 " on path %s",
+		 ftag.segno, path);
 
 	PG_RETURN_VOID();
 }
@@ -166,8 +166,8 @@ test_slru_page_delete(PG_FUNCTION_ARGS)
 	ftag.segno = pageno / SLRU_PAGES_PER_SEGMENT;
 	SlruDeleteSegment(TestSlruCtl, ftag.segno);
 
-	elog(NOTICE, "Called SlruDeleteSegment() for segment %lld",
-		 (long long) ftag.segno);
+	elog(NOTICE, "Called SlruDeleteSegment() for segment %" PRIu64,
+		 ftag.segno);
 
 	PG_RETURN_VOID();
 }
